Make container of auth modal and move related functions to the container
